refactor: extract FileDropOverlay into its own component
